Add a locator button that doesn't change selected routes
I tend to care about the same routes all the time, so the current locate
button isn't useful, since it will select a bunch of routes that aren't
useful to me and possibly clear some that are.

But I often want to zoom and center to my current location.  So this makes
a button that does that but without changing what routes are selected.

I used the bullseye icon for this, since it seems more directly
location-related, and put the current 'locate' behavior behind a bus icon
(since it finds nearby busses. Not entirely accurate or specific, but
not the worst thing ever).

Also changes the zoom level to something closer to how I generally like
to view the map.